+++ rotkehlchen/chain/ethereum/modules/makerdao/constants.py
@@ -10,6 +10,9 @@ RAD = 10**RAD_DIGITS
 CPT_VAULT = 'makerdao vault'
 CPT_DSR = 'makerdao dsr'
 CPT_MIGRATION = 'makerdao migration'
+CPT_SDAI = 'sDAI'
 
 MAKERDAO_LABEL = 'Makerdao'
 MAKERDAO_ICON = 'makerdao.svg'
+SDAI_LABEL = 'sDAI contract'
+SDAI_ICON = 'sdai.svg'