+++ rotkehlchen/tests/unit/decoders/test_main.py
@@ -137,6 +137,7 @@ def test_decoders_initialization(ethereum_transaction_decoder: 'EthereumTransact
         'diva',
         'arbitrum_one',
         'base',
+        'sDAI',
     }